refactor: extract attach and detach buttons into AbstractValidationPage
