+++ vaadin-flow-components-shared-parent/vaadin-flow-components-test-util/src/main/java/com/vaadin/tests/validation/AbstractValidationIT.java
@@ -21,6 +21,8 @@ import com.vaadin.tests.AbstractComponentIT;
 import org.junit.Assert;
 import org.junit.Before;
 
+import static com.vaadin.tests.validation.AbstractValidationPage.ATTACH_FIELD_BUTTON;
+import static com.vaadin.tests.validation.AbstractValidationPage.DETACH_FIELD_BUTTON;
 import static com.vaadin.tests.validation.AbstractValidationPage.SERVER_VALIDITY_STATE;
 import static com.vaadin.tests.validation.AbstractValidationPage.SERVER_VALIDITY_STATE_BUTTON;
 
@@ -87,5 +89,12 @@ public abstract class AbstractValidationIT<T extends TestBenchElement>
         return (Boolean) getCommandExecutor().executeScript(
                 "const field = arguments[0]; const invalid = arguments[1]; return field._shouldSetInvalid(invalid)",
                 testField, invalid);
+    }
+
+    protected void detachAndReattachField() {
+        $("button").id(DETACH_FIELD_BUTTON).click();
+        $("button").id(ATTACH_FIELD_BUTTON).click();
+
+        testField = getTestField();
     }
 }
